Removed changes to User.java file.
Added UserTripData.java entity to encapsulate currentTrip, tripHistory, and username associated with these two things.
