+++ src/main/java/entity/User.java
@@ -1,5 +1,4 @@
 package entity;
-import java.util.List;
 
 /**
  * A simple entity representing a user. Users have a username and password..
@@ -8,8 +7,6 @@ public class User {
 
     private final String name;
     private final String password;
-    private Trip currentTrip;
-    private List<Trip> tripHistory;
 
     /**
      * Creates a new user with the given non-empty name and non-empty password.
@@ -18,7 +15,6 @@ public class User {
      * @throws IllegalArgumentException if the password or name are empty
      */
 
-    // TODO: Implement how user trip info gets
     public User(String name, String password) {
         if ("".equals(name)) {
             throw new IllegalArgumentException("Username cannot be empty");
@@ -37,29 +33,4 @@ public class User {
     public String getPassword() {
         return password;
     }
-
-    public Trip getCurrentTrip() {
-        return currentTrip;
-    }
-
-    public List<Trip> getTripHistory() {
-        return tripHistory;
-    }
-
-    public void completeTrip() {
-        this.tripHistory.add(this.currentTrip);
-        this.currentTrip = null;
-    }
-
-    public void removeCurrentTrip() {
-        this.currentTrip = null;
-    }
-
-    public void createNewTrip(Trip newTrip) {
-        // TODO: Edit This
-        if (this.currentTrip == null) {
-            this.currentTrip = newTrip;
-        }
-    }
-
 }
